Extract dictionary settings tab into module
Split `DictionarySettingsTab` out of `:composeApp` into a new `:dictionary-settings-tab` module, including its Gradle setup, AGENT/CLAUDE notes, detekt baseline, tests, and screenshot assets. Wire the new module back into `composeApp` and settings include list, update CI/workflows (`test.yml`, `screenshots.yml`, `test-changed.sh`) to run its detekt/tests/coverage/screenshots, and update project docs to reflect the new architecture. Also move shared UI test helpers (`unlabelledControls`, `mixedCaseInstalledFont`) into `ui-components` fixtures and adjust related tests/imports.
