+++ composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/dialogs/OptionsDialog.kt
@@ -35,7 +35,8 @@ import org.churchpresenter.app.churchpresenter.centeredOnMainWindow
 import org.churchpresenter.app.churchpresenter.dialogSizeWithin
 import org.churchpresenter.app.churchpresenter.primaryScreenSizeDp
 import org.churchpresenter.core.models.scene.Scene
-import org.churchpresenter.app.churchpresenter.dialogs.tabs.DictionarySettingsTab
+import org.churchpresenter.dictionary.settings.DictionarySettingsTab
+import org.churchpresenter.dictionary.settings.DictionarySettingsTab
 import org.churchpresenter.resources.generated.resources.Res
 import org.churchpresenter.resources.generated.resources.appearance
 import org.churchpresenter.resources.generated.resources.background
