+++ dictionary-settings-tab/screenshots/dictionarySettingsTab/all_hidden.png
