+++ dictionary-settings-tab/screenshots/dictionarySettingsTab/defaults.png
