+++ dictionary-settings-tab/screenshots/dictionarySettingsTab/fades_off.png
