+++ dictionary-settings-tab/screenshots/dictionarySettingsTab/picker_card_background.png
