+++ dictionary-settings-tab/screenshots/dictionarySettingsTab/picker_definition.png
