+++ dictionary-settings-tab/screenshots/dictionarySettingsTab/picker_kjv_usage.png
