+++ dictionary-settings-tab/screenshots/dictionarySettingsTab/picker_word.png
