+++ dictionary-settings-tab/screenshots/dictionarySettingsTab/picker_word_shadow.png
