+++ dictionary-settings-tab/screenshots/dictionarySettingsTab/styled.png
