+++ dictionary-settings-tab/src/test/kotlin/org/churchpresenter/dictionary/settings/DictionarySettingsTabTestSupport.kt
@@ -1,6 +1,6 @@
 @file:OptIn(androidx.compose.ui.test.ExperimentalTestApi::class)
 
-package org.churchpresenter.app.churchpresenter.dialogs.tabs
+package org.churchpresenter.dictionary.settings
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.getValue
@@ -18,15 +18,12 @@ import androidx.compose.ui.test.SemanticsMatcher
 import androidx.compose.ui.test.SemanticsNodeInteraction
 import androidx.compose.ui.test.SemanticsNodeInteractionCollection
 import androidx.compose.ui.test.click
-import androidx.compose.ui.test.hasClickAction
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onRoot
 import androidx.compose.ui.test.performMouseInput
 import androidx.compose.ui.test.performScrollTo
 import androidx.compose.ui.test.runComposeUiTest
 import org.churchpresenter.settings.AppSettings
-import org.churchpresenter.ui.FontCatalog
-import org.churchpresenter.ui.Utils
 import org.churchpresenter.settings.DictionarySettings
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
@@ -122,22 +119,6 @@ internal fun ComposeUiTest.switches(): SemanticsNodeInteractionCollection =
 
 internal fun ComposeUiTest.switch(ordinal: Int): SemanticsNodeInteraction = switches()[ordinal]
 
-/**
- * Every control on the tab with no accessible name at all: no text, no description, no role.
- *
- * A sweep rather than a lookup — the tab's other completeness guards check text and descriptions
- * separately, and a control carrying neither is invisible to both. It used to find the expand arrow
- * inside each font dropdown, which was a bare clickable `Icon`; the font field is one labelled
- * control now, and this is what keeps the next one from going in unnamed.
- */
-internal fun ComposeUiTest.unlabelledControls(): SemanticsNodeInteractionCollection =
-    onAllNodes(
-        hasClickAction() and
-            SemanticsMatcher.keyNotDefined(SemanticsProperties.Text) and
-            SemanticsMatcher.keyNotDefined(SemanticsProperties.EditableText) and
-            SemanticsMatcher.keyNotDefined(SemanticsProperties.ContentDescription) and
-            SemanticsMatcher.keyNotDefined(SemanticsProperties.Role),
-    )
 
 // ── Sliders ─────────────────────────────────────────────────────────────────────────────────────
 
@@ -247,14 +228,3 @@ internal fun ComposeUiTest.assertDurationReads(stored: Float) {
     )
 }
 
-/**
- * An installed font family whose name is not already all-lowercase.
- *
- * `FontSettingsDropdown` seeds its filter with the value it is currently showing and matches
- * `ignoreCase = true`, so parking a field on the lowercased spelling makes the menu offer the
- * properly-cased family — a value the field does not already hold, on any machine's font list.
- */
-internal fun mixedCaseInstalledFont(): String =
-    // Offerable, not merely installed: ".AppleSystemUIFont" is the first mixed-case family on a Mac
-    // and is one of the system-internal faces the picker hides.
-    Utils.getAvailableSystemFonts().first { it != it.lowercase() && !FontCatalog.isHidden(it) }
