+++ dictionary-settings-tab/src/test/kotlin/org/churchpresenter/dictionary/settings/DictionarySettingsTabTextSectionsTest.kt
@@ -1,6 +1,6 @@
 @file:OptIn(androidx.compose.ui.test.ExperimentalTestApi::class)
 
-package org.churchpresenter.app.churchpresenter.dialogs.tabs
+package org.churchpresenter.dictionary.settings
 
 import androidx.compose.ui.test.assertIsOff
 import androidx.compose.ui.test.assertIsOn
